crypto_quant/requirements.txt
# CryptoQuant 依赖

# ==================== 核心依赖 ====================
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# ==================== 数据采集 ====================
requests>=2.26.0
aiohttp>=3.8.0
ccxt>=4.0.0
websocket-client>=1.2.0

# ==================== Qlib (可选，如需原版Qlib功能) ====================
# pyqlib>=0.9.0

# ==================== 机器学习 ====================
lightgbm>=3.3.0
xgboost>=1.5.0
scikit-learn>=1.0.0
joblib>=1.1.0

# ==================== 深度学习 (可选) ====================
# torch>=1.10.0
# tensorflow>=2.7.0

# ==================== 数据存储 ====================
pyarrow>=6.0.0
fastparquet>=0.7.0

# ==================== 配置管理 ====================
pyyaml>=6.0
python-dotenv>=0.19.0

# ==================== 日志 ====================
loguru>=0.5.0

# ==================== 可视化 ====================
matplotlib>=3.5.0
plotly>=5.4.0
seaborn>=0.11.0

# ==================== 异步 ====================
asyncio-throttle>=1.0.0
aiolimiter>=1.0.0

# ==================== 工具 ====================
tqdm>=4.62.0
click>=8.0.0
tabulate>=0.8.9

# ==================== 测试 ====================
pytest>=6.2.0
pytest-asyncio>=0.16.0
pytest-cov>=3.0.0

# ==================== 开发 (可选) ====================
# black>=21.0
# flake8>=4.0.0
# mypy>=0.910
# pre-commit>=2.15.0

# ==================== Jupyter (可选) ====================
# jupyter>=1.0.0
# jupyterlab>=3.2.0
# ipywidgets>=7.6.0
